+++ src/35_pt/StopArea.cpp
@@ -216,12 +216,12 @@ namespace synthese
 					}
 					result.insert(
 						it.second->getProjectedPoint().getRoadChunk()->getFromCrossing(),
-						VertexAccess(minutes(it.second->getProjectedPoint().getMetricOffset() / 50), it.second->getProjectedPoint().getMetricOffset())
+						VertexAccess(minutes(static_cast<long>(it.second->getProjectedPoint().getMetricOffset() / 50)), it.second->getProjectedPoint().getMetricOffset())
 					);
 					result.insert(
 						it.second->getProjectedPoint().getRoadChunk()->getReverseRoadChunk()->getFromCrossing(),
 						VertexAccess(
-							minutes((it.second->getProjectedPoint().getRoadChunk()->getEndMetricOffset() - it.second->getProjectedPoint().getRoadChunk()->getMetricOffset() - it.second->getProjectedPoint().getMetricOffset()) / 50),
+							minutes(static_cast<long>((it.second->getProjectedPoint().getRoadChunk()->getEndMetricOffset() - it.second->getProjectedPoint().getRoadChunk()->getMetricOffset() - it.second->getProjectedPoint().getMetricOffset()) / 50)),
 							it.second->getProjectedPoint().getRoadChunk()->getEndMetricOffset() - it.second->getProjectedPoint().getRoadChunk()->getMetricOffset() - it.second->getProjectedPoint().getMetricOffset()
 					)	);
 				}
